Completed iris test migration
